feat: added filter params

+++ client/components/App.js
@@ -1,12 +1,15 @@
 import React from 'react';
+import Filter from './Filter';
 import SingleTodo from './SingleTodo';
 
 const App = () => {
 	let [change, setChange] = React.useState(false);
+	let [filter, setFilter] = React.useState('all');
 	let [todos, setTodos] = React.useState([]);
 	let [input, setInput] = React.useState('');
 	let [allTrue, setAllTrue] = React.useState(true);
 	const mounted = React.useRef(true);
+	let count = 0;
 
 	React.useEffect(async () => {
 		mounted.current = true;
@@ -15,7 +18,6 @@ const App = () => {
 		const response = await (
 			await fetch('https://todo-halo.herokuapp.com/todos')
 		).json();
-		console.log(response);
 		if (mounted.current) setTodos(response.sort((a, b) => a.id - b.id));
 		if (change) setChange(false);
 		if (!mounted.current) mounted.current = true;
@@ -45,10 +47,8 @@ const App = () => {
 		return res.data;
 	};
 	const markAll = () => {
-		console.log(allTrue, 'before');
 		todos.forEach((each) => changeDone(each, allTrue));
 		setAllTrue(!allTrue);
-		console.log(allTrue, 'after');
 	};
 	const deleteItem = async (id) => {
 		const res = await fetch(`https://todo-halo.herokuapp.com/todos/${id}`, {
@@ -56,6 +56,11 @@ const App = () => {
 			headers: { 'Content-Type': 'application/json' },
 		});
 		if (res) setChange(true);
+	};
+	const deleteAll = () => {
+		todos.forEach((each) => {
+			if (each.isDone) return deleteItem(each.id);
+		});
 	};
 	return (
 		<>
@@ -80,18 +85,31 @@ const App = () => {
 						></input>
 					</form>
 				</section>
-				{todos && todos.length
-					? todos.map((each) => (
-							<SingleTodo
-								content={each.content}
-								isDone={each.isDone}
-								key={each.id}
-								id={each.id}
-								handleClick={changeDone}
-								delete={deleteItem}
-							/>
-					  ))
-					: null}
+				{todos && todos.length ? (
+					<>
+						{todos.map((each) => {
+							if (filter === 'all' || filter == `${each.isDone}`) {
+								count += 1;
+								return (
+									<SingleTodo
+										content={each.content}
+										isDone={each.isDone}
+										key={each.id}
+										id={each.id}
+										handleClick={changeDone}
+										delete={deleteItem}
+									/>
+								);
+							}
+						})}
+						<Filter
+							deleteAll={deleteAll}
+							count={count}
+							setFilter={setFilter}
+							filter={filter}
+						/>
+					</>
+				) : null}
 			</section>
 		</>
 	);
